+++ dramalyzer.py
@@ -3,31 +3,29 @@
 #
 # dramavis by frank fischer (@umblaetterer) & christopher kittel (@chris_kittel)
 
+import os
+import csv
+from itertools import chain, zip_longest
+from collections import Counter
+import logging
+import numpy as np
+import pandas as pd
+import networkx as nx
+from scipy import stats
+import statsmodels.api as sm
+import statsmodels.formula.api as sm_formula
+from tqdm import tqdm
+
+from linacorpus import LinaCorpus, Lina
+from dramaplotter import plotGraph
+
+
 __author__ = "Christopher Kittel <web at christopherkittel.eu>, Frank Fischer <ffischer at hse.ru>"
 __copyright__ = "Copyright 2017"
 __license__ = "MIT"
 __version__ = "0.4 (beta)"
 __maintainer__ = "Frank Fischer <ffischer at hse.ru>"
 __status__ = "Development" # 'Development', 'Production' or 'Prototype'
-
-import os
-import glob
-import pandas as pd
-import networkx as nx
-import csv
-from lxml import etree
-from itertools import chain, zip_longest
-from collections import Counter
-import argparse
-from superposter import plotGraph, plot_superposter
-import logging
-import numpy as np
-from scipy import stats, optimize
-import statsmodels.api as sm
-import statsmodels.formula.api as sm_formula
-from tqdm import tqdm
-
-from linacorpus import LinaCorpus, Lina
 
 
 class CorpusAnalyzer(LinaCorpus):
