+++ superposter.py
@@ -9,46 +9,6 @@ import matplotlib.gridspec as gridspec
 import os
 import math
 
-def plotGraph(G, figsize=(8, 8), filename=None):
-    """
-    Plots an individual graph, node size by degree centrality,
-    edge size by edge weight.
-    """
-    labels = {n:n for n in G.nodes()}
-
-    try:
-        # for networks with only one node
-        d = nx.degree_centrality(G)
-        nodesize = [v * 250 for v in d.values()]
-    except:
-        nodesize = [1 * 250 for n in G.nodes()]
-
-    layout=nx.spring_layout
-    pos=layout(G)
-
-    plt.figure(figsize=figsize)
-    plt.subplots_adjust(left=0,right=1,bottom=0,top=0.95,wspace=0.01,hspace=0.01)
-
-    # nodes
-    nx.draw_networkx_nodes(G,pos,
-                            nodelist=G.nodes(),
-                            node_color="steelblue",
-                            node_size=nodesize,
-                            alpha=0.8)
-    try:
-        weights = [G[u][v]['weight'] for u,v in G.edges()]
-    except:
-        weights = [1 for u,v in G.edges()]
-    nx.draw_networkx_edges(G,pos,
-                           with_labels=False,
-                           edge_color="grey",
-                           width=weights
-                        )
-
-    if G.order() < 1000:
-        nx.draw_networkx_labels(G,pos, labels)
-    plt.savefig(filename)
-    plt.close("all")
 
 def plot_superposter(corpus, outputdir, debug=False):
     """
@@ -70,8 +30,10 @@ def plot_superposter(corpus, outputdir, debug=False):
     bottom, height = .25, .5
     right = left + width
     top = bottom + height
-    dramas = {drama.ID:drama for drama in corpus.read_dramas(metrics=False)}
-    id2date = {drama.ID:drama.metadata.get("date_definite") for drama in corpus.read_dramas(metrics=False)}
+    dramas = {drama.ID:drama
+              for drama in corpus.read_dramas(metrics=False)}
+    id2date = {drama.ID:drama.metadata.get("date_definite")
+               for drama in corpus.read_dramas(metrics=False)}
     if debug:
         print(id2date)
